refactorings, work on nlu data, config uncommenting
