+++ backend/assistant_core/.gitignore
@@ -1,1 +1,2 @@
-actions/__pycache__
+__pycache__
+**/__pycache__
