+++ backend/assistant_core/actions/actions.py
@@ -1,18 +1,12 @@
-# This files contains your custom actions which can be used to run
-# custom Python code.
-#
 # See this guide on how to implement these action:
 # https://rasa.com/docs/rasa/custom-actions
-
-
-# This is a simple example for a custom action which utters "Hello World!"
 
 from typing import Any, Text, Dict, List
 
 from rasa_sdk import Action, Tracker
 from rasa_sdk.executor import CollectingDispatcher
 # from rasa_sdk.events import SlotSet
-from wx import wx_city
+from src.wx import wx_city
 
 
 class Weather(Action):
