+++ backend/assistant_core/domain.yml
@@ -8,6 +8,7 @@ intents:
   - mood_great
   - mood_unhappy
   - bot_challenge
+  - capabilities
   - book_a_ticket
   - buy_groceries
   - wants_a_joke
