+++ backend/assistant_core/models/20211108-212059.tar.gz
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:46008d6208d5af6be73cd2fff984669dc593d84fb2d742e737639969c7dcb482
+size 26988140
