+++ backend/assistant_core/src/wx.py
@@ -16,7 +16,7 @@ def wx_city(city, use_API=False):
             lat, lon = 47.49, 19.04
         if city == "Debrecen":
             lat, lon = 47.53, 21.62
-        if city == "Szeged":
+        else: # city == "Szeged":
             lat, lon = 46.25, 20.14
          
         API_key = "0f7822807230d53f4377f28275dda1b4"
@@ -35,7 +35,7 @@ def wx_city(city, use_API=False):
         if city == "Debrecen":
             wx_data={'lat': -41.2, 'lon': 174.8, 'timezone': 'Pacific/Auckland', 'timezone_offset': 46800, 'current': {'dt': 1607204356, 'sunrise': 1607186509, 'sunset': 1607240525, 'temp': 17.56, 'feels_like': 14.68, 'pressure': 1011, 'humidity': 55, 'dew_point': 8.43, 'uvi': 7.43, 'clouds': 0, 'visibility': 10000, 'wind_speed': 3.6, 'wind_deg': 60, 'weather': [{'id': 800, 'main': 'Clear', 'description': 'clear sky', 'icon': '01d'}]}, 'daily': [{'dt': 1607212800, 'sunrise': 1607186509, 'sunset': 1607240525, 'temp': {'day': 15.36, 'min': 10.64, 'max': 17.56, 'night': 13.39, 'eve': 13.41, 'morn': 11.46}, 'feels_like': {'day': 10.13, 'night': 6.02, 'eve': 8.1, 'morn': 10.97}, 'pressure': 1011, 'humidity': 65, 'dew_point': 8.82, 'wind_speed': 7.09, 'wind_deg': 355, 'weather': [{'id': 500, 'main': 'Rain', 'description': 'light rain', 'icon': '10d'}], 'clouds': 31, 'pop': 0.48, 'rain': 1.21, 'uvi': 10.7}, {'dt': 1607299200, 'sunrise': 1607272903, 'sunset': 1607326982, 'temp': {'day': 15.84, 'min': 13.54, 'max': 16.43, 'night': 15.82, 'eve': 16.07, 'morn': 14.47}, 'feels_like': {'day': 8.5, 'night': 8.55, 'eve': 9.7, 'morn': 7.18}, 'pressure': 1006, 'humidity': 97, 'dew_point': 15.48, 'wind_speed': 12.98, 'wind_deg': 325, 'weather': [{'id': 501, 'main': 'Rain', 'description': 'moderate rain', 'icon': '10d'}], 'clouds': 100, 'pop': 0.83, 'rain': 6.12, 'uvi': 5.85}, {'dt': 1607385600, 'sunrise': 1607359299, 'sunset': 1607413437, 'temp': {'day': 16.35, 'min': 13.93, 'max': 17.03, 'night': 14.56, 'eve': 14.18, 'morn': 15.95}, 'feels_like': {'day': 10.45, 'night': 10.28, 'eve': 10, 'morn': 9.88}, 'pressure': 1007, 'humidity': 90, 'dew_point': 14.84, 'wind_speed': 10.59, 'wind_deg': 329, 'weather': [{'id': 501, 'main': 'Rain', 'description': 'moderate rain', 'icon': '10d'}], 'clouds': 100, 'pop': 0.73, 'rain': 5.53, 'uvi': 10.11}, {'dt': 1607472000, 'sunrise': 1607445698, 'sunset': 1607499890, 'temp': {'day': 15.99, 'min': 14.61, 'max': 16.02, 'night': 15.39, 'eve': 15.52, 'morn': 15.81}, 'feels_like': {'day': 11.66, 'night': 12.76, 'eve': 11.62, 'morn': 10.4}, 'pressure': 1007, 'humidity': 99, 'dew_point': 15.93, 'wind_speed': 8.94, 'wind_deg': 336, 'weather': [{'id': 501, 'main': 'Rain', 'description': 'moderate rain', 'icon': '10d'}], 'clouds': 100, 'pop': 0.99, 'rain': 11.13, 'uvi': 6.01}, {'dt': 1607558400, 'sunrise': 1607532098, 'sunset': 1607586342, 'temp': {'day': 13.44, 'min': 13.44, 'max': 16.97, 'night': 13.68, 'eve': 15.77, 'morn': 14.82}, 'feels_like': {'day': 12.87, 'night': 8.2, 'eve': 12.08, 'morn': 12.19}, 'pressure': 1006, 'humidity': 97, 'dew_point': 13.08, 'wind_speed': 2.13, 'wind_deg': 118, 'weather': [{'id': 502, 'main': 'Rain', 'description': 'heavy intensity rain', 'icon': '10d'}], 'clouds': 89, 'pop': 1, 'rain': 18.34, 'uvi': 10.48}, {'dt': 1607644800, 'sunrise': 1607618502, 'sunset': 1607672792, 'temp': {'day': 15.25, 'min': 13.55, 'max': 15.3, 'night': 13.83, 'eve': 14.43, 'morn': 13.93}, 'feels_like': {'day': 8.08, 'night': 7.99, 'eve': 8.28, 'morn': 7.07}, 'pressure': 1012, 'humidity': 82, 'dew_point': 12.28, 'wind_speed': 11.21, 'wind_deg': 350, 'weather': [{'id': 804, 'main': 'Clouds', 'description': 'overcast clouds', 'icon': '04d'}], 'clouds': 93, 'pop': 0.04, 'uvi': 11}, {'dt': 1607731200, 'sunrise': 1607704907, 'sunset': 1607759241, 'temp': {'day': 15.77, 'min': 13.32, 'max': 15.82, 'night': 13.34, 'eve': 14.61, 'morn': 13.46}, 'feels_like': {'day': 14.51, 'night': 11.96, 'eve': 13.93, 'morn': 10.38}, 'pressure': 1016, 'humidity': 77, 'dew_point': 11.88, 'wind_speed': 2.58, 'wind_deg': 332, 'weather': [{'id': 500, 'main': 'Rain', 'description': 'light rain', 'icon': '10d'}], 'clouds': 100, 'pop': 0.32, 'rain': 0.25, 'uvi': 11}, {'dt': 1607817600, 'sunrise': 1607791315, 'sunset': 1607845688, 'temp': {'day': 13.18, 'min': 12.46, 'max': 13.72, 'night': 12.46, 'eve': 12.86, 'morn': 12.73}, 'feels_like': {'day': 10.39, 'night': 8.47, 'eve': 9.26, 'morn': 10.82}, 'pressure': 1017, 'humidity': 95, 'dew_point': 12.55, 'wind_speed': 5.05, 'wind_deg': 157, 'weather': [{'id': 501, 'main': 'Rain', 'description': 'moderate rain', 'icon': '10d'}], 'clouds': 100, 'pop': 0.88, 'rain': 7.37, 'uvi': 11}]}
 
-        if city == "Szeged":
+        else:# city == "Szeged":
             wx_data={'lat': -43.5, 'lon': 172.6, 'timezone': 'Pacific/Auckland', 'timezone_offset': 46800, 'current': {'dt': 1607206141, 'sunrise': 1607186569, 'sunset': 1607241522, 'temp': 16, 'feels_like': 11.67, 'pressure': 1009, 'humidity': 67, 'dew_point': 9.88, 'uvi': 6.27, 'clouds': 32, 'visibility': 10000, 'wind_speed': 6.2, 'wind_deg': 40, 'weather': [{'id': 802, 'main': 'Clouds', 'description': 'scattered clouds', 'icon': '03d'}]}, 'daily': [{'dt': 1607212800, 'sunrise': 1607186569, 'sunset': 1607241522, 'temp': {'day': 14.75, 'min': 8.71, 'max': 16.49, 'night': 14.53, 'eve': 16.49, 'morn': 9.84}, 'feels_like': {'day': 9.49, 'night': 7.12, 'eve': 10.55, 'morn': 7.96}, 'pressure': 1008, 'humidity': 69, 'dew_point': 9.13, 'wind_speed': 7.25, 'wind_deg': 71, 'weather': [{'id': 803, 'main': 'Clouds', 'description': 'broken clouds', 'icon': '04d'}], 'clouds': 51, 'pop': 0.79, 'uvi': 9.28}, {'dt': 1607299200, 'sunrise': 1607272960, 'sunset': 1607327981, 'temp': {'day': 23.34, 'min': 13.24, 'max': 24.35, 'night': 16.35, 'eve': 21.87, 'morn': 16.2}, 'feels_like': {'day': 22.1, 'night': 13.38, 'eve': 19.16, 'morn': 13.63}, 'pressure': 1000, 'humidity': 43, 'dew_point': 10.29, 'wind_speed': 1.85, 'wind_deg': 353, 'weather': [{'id': 804, 'main': 'Clouds', 'description': 'overcast clouds', 'icon': '04d'}], 'clouds': 98, 'pop': 0.06, 'uvi': 9.77}, {'dt': 1607385600, 'sunrise': 1607359353, 'sunset': 1607414439, 'temp': {'day': 24.32, 'min': 12.24, 'max': 24.97, 'night': 16.17, 'eve': 19.02, 'morn': 13.99}, 'feels_like': {'day': 21.62, 'night': 13.8, 'eve': 16.42, 'morn': 13.84}, 'pressure': 1002, 'humidity': 41, 'dew_point': 10.31, 'wind_speed': 4.01, 'wind_deg': 287, 'weather': [{'id': 804, 'main': 'Clouds', 'description': 'overcast clouds', 'icon': '04d'}], 'clouds': 100, 'pop': 0, 'uvi': 8.64}, {'dt': 1607472000, 'sunrise': 1607445749, 'sunset': 1607500895, 'temp': {'day': 19.58, 'min': 12.81, 'max': 19.58, 'night': 13.11, 'eve': 15.64, 'morn': 12.81}, 'feels_like': {'day': 18.33, 'night': 12.16, 'eve': 13.21, 'morn': 9.65}, 'pressure': 1006, 'humidity': 56, 'dew_point': 10.78, 'wind_speed': 2.07, 'wind_deg': 127, 'weather': [{'id': 804, 'main': 'Clouds', 'description': 'overcast clouds', 'icon': '04d'}], 'clouds': 100, 'pop': 0, 'uvi': 8.84}, {'dt': 1607558400, 'sunrise': 1607532147, 'sunset': 1607587350, 'temp': {'day': 19.65, 'min': 11.67, 'max': 19.69, 'night': 12.79, 'eve': 16.31, 'morn': 13.36}, 'feels_like': {'day': 17.21, 'night': 10.89, 'eve': 13.66, 'morn': 10.17}, 'pressure': 1006, 'humidity': 60, 'dew_point': 11.72, 'wind_speed': 4.23, 'wind_deg': 203, 'weather': [{'id': 803, 'main': 'Clouds', 'description': 'broken clouds', 'icon': '04d'}], 'clouds': 54, 'pop': 0, 'uvi': 10.55}, {'dt': 1607644800, 'sunrise': 1607618548, 'sunset': 1607673802, 'temp': {'day': 21.76, 'min': 11.48, 'max': 22.35, 'night': 14.14, 'eve': 18.31, 'morn': 12.33}, 'feels_like': {'day': 20.92, 'night': 10.02, 'eve': 16.57, 'morn': 11.24}, 'pressure': 1009, 'humidity': 45, 'dew_point': 9.51, 'wind_speed': 1, 'wind_deg': 332, 'weather': [{'id': 804, 'main': 'Clouds', 'description': 'overcast clouds', 'icon': '04d'}], 'clouds': 100, 'pop': 0, 'uvi': 11}, {'dt': 1607731200, 'sunrise': 1607704951, 'sunset': 1607760253, 'temp': {'day': 15.47, 'min': 12.76, 'max': 15.49, 'night': 12.85, 'eve': 13.29, 'morn': 13.15}, 'feels_like': {'day': 13.02, 'night': 10.44, 'eve': 9.85, 'morn': 11.84}, 'pressure': 1017, 'humidity': 79, 'dew_point': 11.87, 'wind_speed': 4.32, 'wind_deg': 96, 'weather': [{'id': 500, 'main': 'Rain', 'description': 'light rain', 'icon': '10d'}], 'clouds': 98, 'pop': 0.3, 'rain': 0.3, 'uvi': 11}, {'dt': 1607817600, 'sunrise': 1607791357, 'sunset': 1607846702, 'temp': {'day': 12.21, 'min': 11.08, 'max': 12.84, 'night': 11.08, 'eve': 11.85, 'morn': 12.05}, 'feels_like': {'day': 9.79, 'night': 9.99, 'eve': 9.97, 'morn': 10}, 'pressure': 1019, 'humidity': 80, 'dew_point': 9.04, 'wind_speed': 3.1, 'wind_deg': 97, 'weather': [{'id': 500, 'main': 'Rain', 'description': 'light rain', 'icon': '10d'}], 'clouds': 100, 'pop': 0.62, 'rain': 1.73, 'uvi': 11}]}
 
     return wx_data
